requirements.txt
streamlit
pandas
plotly
sqlalchemy
psycopg2-binary
google-generativeai
python-dotenv
